requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.8.0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==5.2.0
psycopg2==2.9.1
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
typing-extensions==3.10.0.2
tzlocal==2.1
yarl==1.7.0
